+++ src/xai_sdk/aio/client.py
@@ -13,7 +13,7 @@ from ..interceptors import (
     UnaryUnaryAuthAioInterceptor,
     UnaryUnaryTimeoutAioInterceptor,
 )
-from . import auth, batch, chat, collections, files, image, models, tokenizer, video
+from . import auth, batch, chat, collections, files, image, models, tokenizer, tts, video
 
 
 class Client(BaseClient):
@@ -27,6 +27,7 @@ class Client(BaseClient):
     image: "image.Client"
     models: "models.Client"
     tokenize: "tokenizer.Client"
+    tts: "tts.Client"
     video: "video.Client"
 
     def _init(
@@ -74,6 +75,13 @@ class Client(BaseClient):
         self.image = image.Client(self._api_channel)
         self.models = models.Client(self._api_channel)
         self.tokenize = tokenizer.Client(self._api_channel)
+        self.tts = tts.Client(
+            api_key,
+            api_host,
+            timeout=timeout,
+            metadata=metadata,
+            use_insecure_channel=use_insecure_channel,
+        )
         self.video = video.Client(self._api_channel)
 
     def _make_grpc_channel(